requirements.txt
tk
pyautogui
gtts
Pillow
pygments
playsound==1.2.2